refactor(effects/combat): Los effects actuales corresponden completamente a habilidades de combate, por ende lo normal seria tenerlos en 'view'
- Al estar los effects en view se respeta el MVC.
- Falta destripar cada effects del view para sacarle el controlador y el modelo.
