+++ roguelike_project/systems/combat/view/effects/particles/explosions/fire.py
@@ -1,6 +1,6 @@
 import random
 import math
-from roguelike_project.systems.effects.particles.particle import Particle
+from roguelike_project.systems.combat.view.effects.particles.particle import Particle
 
 class FireExplosion:
     def __init__(self, x, y, particle_count=100):
