+++ roguelike_project/systems/combat/view/effects/particles/spells/dash_bounce.py
@@ -1,6 +1,6 @@
 import math
 import random
-from roguelike_project.systems.effects.particles.particle import Particle
+from roguelike_project.systems.combat.view.effects.particles.particle import Particle
 
 class DashBounce:
     def __init__(self, x, y):
